+++ boolean-logic-ultimate/Invert.js
@@ -3,19 +3,39 @@ module.exports = function(RED) {
         RED.nodes.createNode(this,config);
 		this.config = config;
 		var node = this;
-		var NodeHelper = require('./NodeHelper.js');
-		var h = new NodeHelper( node );
+
 		
         this.on('input', function(msg) {
 			var topic = msg.topic;
 			var payload = msg.payload;
 			
-			if( topic !== undefined && payload !== undefined ) {
-				h.SetResult( !h.ToBoolean( payload ), topic );
+			if (topic !== undefined && payload !== undefined) {
+				node.status( {fill:  "green" ,shape: "dot" ,text: !ToBoolean(payload)});
+				return({ topic: topic, payload: !ToBoolean(payload) });
 			}
         });
 		
-		h.DisplayUnkownStatus();
+		node.status( {fill:  "grey" ,shape: "dot" ,text: payload});
+
+		function ToBoolean( value ) {
+			var res = false;
+	
+			if (typeof value === 'boolean') {
+				res = value;
+			} 
+			else if( typeof value === 'number' || typeof value === 'string' ) {
+				// Is it formated as a decimal number?
+				if( decimal.test( value ) ) {
+					var v = parseFloat( value );
+					res = v != 0;
+				}
+				else {
+					res = value.toLowerCase() === "true";
+				}
+			}
+			
+			return res;
+		};
     }