+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # node-red-contrib-boolean-logic-ultimate
 <p>
+<b>Version 0.0.2</b><br/>
+- Fixed persistent state deletion upon node update/delete<br/>
+</p>
+<p>
 <b>Version 0.0.1</b><br/>
 - Initial release<br/>
 </p>